fix: address PR #488 review + CI iOS test (PUL-22)
- iOS: split BudgetDetailsView and BudgetDetailsProjector into sibling
  extension files (routing, previews, sections) so every BudgetDetails file
  stays under the 350-LOC architecture invariant — the failing CI test.
- backend: resolve the next-month budget with pure month/year arithmetic.
  date-fns addMonths applied local-time calendar math to a UTC date, so on a
  DST host it could target the source month. Document the 409 responses with
  @ApiConflictResponse on both postpone endpoints.
- webapp: expose postpone on free-transaction rows in the desktop table.
  They render through actions-cell but the action was gated to budget lines;
  route the table's postpone output by item type.
