+++ ios/Pulpe/Features/Budgets/BudgetDetails/BudgetDetailsView+Previews.swift
@@ -1,0 +1,28 @@
+import SwiftUI
+import TipKit
+
+// MARK: - Previews
+
+#Preview {
+    NavigationStack {
+        BudgetDetailsView(budgetId: "test")
+    }
+    .environment(AppState())
+    .environment(BudgetDetailsRouter())
+    .environment(UserSettingsStore())
+    .environment(BudgetListStore())
+    .environment(DashboardStore())
+}
+
+#Preview("Gestures Tip") {
+    List {
+        Section("Dépenses") {
+            TipView(ProductTips.gestures)
+            Text("Courses alimentaires")
+        }
+    }
+    .listStyle(.insetGrouped)
+    .scrollContentBackground(.hidden)
+    .pulpeBackground()
+    .task { try? Tips.resetDatastore() }
+}
