+++ ios/Pulpe/Features/Budgets/BudgetDetails/Projection/BudgetDetailsProjector.swift
@@ -187,65 +187,6 @@ final class BudgetDetailsProjector {
         )
     }
 
-    private static func makeSections(
-        dataStore: BudgetDataStore,
-        filtersStore: FiltersStore,
-        syncStore: SyncStateStore,
-        searchText: String,
-        consumptionByLineId: [String: BudgetFormulas.Consumption]
-    ) -> [BudgetDetailsScreenState.Section] {
-        let syncing = syncStore.syncingBudgetLineIds
-        // Line ids carrying at least one allocated transaction — folds the
-        // PUL-22 CA7 "no allocated tx" check into per-line eligibility below.
-        let allocatedLineIds = Set(dataStore.transactions.compactMap(\.budgetLineId))
-        var sections: [BudgetDetailsScreenState.Section] = []
-        let displayed = filtersStore.displayedSections(for: dataStore.budgetLines)
-        sections.reserveCapacity(displayed.count)
-        for section in displayed {
-            let searchFiltered = filtersStore.filteredLines(
-                section.items,
-                searchText: searchText,
-                transactions: dataStore.transactions
-            )
-            guard !searchFiltered.isEmpty else { continue }
-            let items = searchFiltered.map { line in
-                BudgetDetailsScreenState.LineItem(
-                    line: line,
-                    consumption: consumptionByLineId[line.id] ?? zeroConsumption(for: line),
-                    isSyncing: syncing.contains(line.id),
-                    isPostponeEligible: line.isPostponeEligible(
-                        hasAllocatedTransactions: allocatedLineIds.contains(line.id)
-                    )
-                )
-            }
-            sections.append(
-                BudgetDetailsScreenState.Section(kind: section.kind, items: items)
-            )
-        }
-        return sections
-    }
-
-    private static func makeFreeItems(
-        dataStore: BudgetDataStore,
-        filtersStore: FiltersStore,
-        syncStore: SyncStateStore,
-        searchText: String
-    ) -> [BudgetDetailsScreenState.FreeTransactionItem] {
-        let syncing = syncStore.syncingTransactionIds
-        return filtersStore.combinedFilteredFreeTransactions(
-            dataStore.freeTransactions,
-            searchText: searchText
-        ).map { tx in
-            BudgetDetailsScreenState.FreeTransactionItem(
-                transaction: tx,
-                isSyncing: syncing.contains(tx.id),
-                // Free transactions are unallocated by construction; eligibility
-                // reduces to "unchecked", mirroring `LineItem.isPostponeEligible`.
-                isPostponeEligible: tx.checkedAt == nil
-            )
-        }
-    }
-
     private static func makeRollover(
         dataStore: BudgetDataStore
     ) -> BudgetDetailsScreenState.RolloverInfo? {
@@ -345,15 +286,5 @@ final class BudgetDetailsProjector {
             hasher.combine(tx.isChecked)
         }
         return hasher.finalize()
-    }
-
-    private static func zeroConsumption(
-        for line: BudgetLine
-    ) -> BudgetFormulas.Consumption {
-        BudgetFormulas.Consumption(
-            allocated: 0,
-            available: line.amount,
-            percentage: 0
-        )
     }
 }
